backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
google-generativeai
googlemaps
aiohttp
httpx
